feat: update frontend for issuers
