--- frontend/src/components/AppButton.vue
@@ -1,49 +1,0 @@
-<template>
-  <div>
-    <button>
-      <slot />
-    </button>
-  </div>
-</template>
-
-<style scoped>
-button {
-  background: none;
-  border: none;
-  color: rgba(0, 125, 255);
-  padding: 0;
-  font-size: inherit;
-  font-weight: 800;
-  font-family: inherit;
-  text-transform: uppercase;
-  margin-top: 0.5rem;
-  cursor: pointer;
-  transition: opacity 0.1s;
-  letter-spacing: 0.03em;
-  transition: color 0.25s;
-  display: inline-flex;
-  align-items: center;
-}
-button:focus {
-  opacity: 0.85;
-  outline: none;
-}
-button:active {
-  opacity: 0.65;
-}
-.button__label {
-  display: inline-flex;
-  align-items: center;
-}
-.button__label__icon {
-  height: 1em;
-  width: 1em;
-  margin: 0 0.5em 0 0.5em;
-  fill: rgba(0, 0, 0, 0.25);
-  animation: rotate linear 4s infinite;
-}
-.button.button__valid__false {
-  color: rgba(0, 0, 0, 0.25);
-  cursor: not-allowed;
-}
-</style>
