--- frontend/src/components/AppRadioItem.vue
@@ -1,47 +1,0 @@
-<template>
-  <div>
-    <button class="button">{{ value }}</button>
-  </div>
-</template>
-
-<style scoped>
-.button {
-  box-shadow: inset 0 0 0 1px rgba(0, 0, 0, 0.1);
-  padding: 0.75rem 1rem;
-  margin-bottom: 0.5rem;
-  border-radius: 6px;
-  user-select: none;
-  cursor: pointer;
-  transition: all 0.1s;
-  font-weight: 500;
-  outline: none;
-  border: none;
-  background: rgba(0, 0, 0, 0.01);
-  width: 100%;
-  font-family: inherit;
-  text-align: left;
-  font-size: 1rem;
-  font-weight: inherit;
-}
-.button:hover {
-  box-shadow: inset 0 0 0 1px rgba(0, 0, 0, 0.1),
-    0 1px 5px -1px rgba(0, 0, 0, 0.1);
-}
-.button:focus {
-  box-shadow: inset 0 0 0 1px rgba(0, 0, 0, 0.2),
-    0 1px 5px -1px rgba(0, 0, 0, 0.1);
-}
-.button:active {
-  box-shadow: inset 0 0 0 1px rgba(0, 0, 0, 0.07);
-  color: rgba(0, 0, 0, 0.7);
-  transform: scale(0.998);
-}
-</style>
-
-<script>
-export default {
-  props: {
-    value: "",
-  },
-};
-</script>
