--- frontend/src/components/AppText.vue
@@ -1,37 +1,0 @@
-<template>
-  <div>
-    <div :class="[`${type}`]">
-      <slot />
-    </div>
-  </div>
-</template>
-
-<style scoped>
-.h1 {
-  font-size: 2rem;
-  font-weight: 800;
-  letter-spacing: 0.03em;
-  margin-bottom: 2rem;
-}
-.h2 {
-  font-weight: 800;
-  text-transform: uppercase;
-  letter-spacing: 0.05em;
-  margin-bottom: 0.5rem;
-  margin-top: 2rem;
-}
-.subtitle {
-  color: rgba(0, 0, 0, 0.5);
-  font-size: 0.85rem;
-}
-</style>
-
-<script>
-export default {
-  props: {
-    type: {
-      default: "p1",
-    },
-  },
-};
-</script>
