--- frontend/src/components/TypeList.vue
@@ -1,17 +1,0 @@
-<template>
-  <div>
-    <div class="type" v-for="type in typeList" :key="type.type">
-      <type-item :value="type" />
-    </div>
-  </div>
-</template>
-
-<script>
-export default {
-  computed: {
-    typeList() {
-      return this.$store.state.app.types || [];
-    },
-  },
-};
-</script>
