--- frontend/src/store/app.js
@@ -1,5 +1,0 @@
-module.exports = {
-  types: [
-    // this line is used by starport scaffolding
-  ],
-};
