--- frontend/src/views/Issuer.vue
@@ -1,7 +1,0 @@
-<template>
-  <div>
-    <app-layout>
-      <issuer-list />
-    </app-layout>
-  </div>
-</template>
