--- frontend/src/views/Poa.vue
@@ -1,7 +1,0 @@
-<template>
-  <div>
-    <app-layout>
-      <poa-validator-list />
-    </app-layout>
-  </div>
-</template>
