+++ vue/.env
@@ -1,0 +1,4 @@
+VUE_APP_ADDRESS_PREFIX=cosmos
+VUE_APP_API_COSMOS=http://127.0.0.1:1317
+VUE_APP_API_TENDERMINT=http://127.0.0.1:26657
+VUE_APP_API_WS=ws://127.0.0.1:26657/websocket
