+++ vue/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "cbdc",
+  "name": "frontend",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -8,11 +8,8 @@
     "build": "vue-cli-service build"
   },
   "dependencies": {
-    "@cosmjs/launchpad": "^0.22.0-alpha.0",
-    "axios": "^0.19.2",
-    "bip39": "^3.0.2",
+    "@tendermint/vue": "0.1.1",
     "core-js": "^3.6.5",
-    "lodash": "^4.17.19",
     "vue": "^2.6.11",
     "vue-material": "^1.0.0-beta-15",
     "vue-router": "^3.2.0",
@@ -23,6 +20,7 @@
     "@vue/cli-plugin-router": "^4.4.0",
     "@vue/cli-plugin-vuex": "^4.4.6",
     "@vue/cli-service": "^4.4.0",
+    "prettier": "2.1.2",
     "vue-template-compiler": "^2.6.11"
   }
 }
