+++ vue/public/favicon.ico
