+++ vue/src/main.js
@@ -1,0 +1,21 @@
+import Vue from "vue";
+import App from "./App.vue";
+import router from "./router";
+import store from "./store";
+
+import VueMaterial from "vue-material";
+import "vue-material/dist/vue-material.min.css";
+import "vue-material/dist/theme/default.css";
+
+import IssuerList from "./components/IssuerList.vue";
+
+Vue.use(VueMaterial);
+Vue.component("issuer-list", IssuerList);
+
+Vue.config.productionTip = false;
+
+new Vue({
+  router,
+  store,
+  render: (h) => h(App),
+}).$mount("#app");
