+++ vue/src/views/Index.vue
@@ -1,0 +1,18 @@
+<template>
+  <div>
+    <div class="sp-container">
+      <sp-sign-in />
+      <sp-bank-balances />
+      <sp-token-send />
+      <!-- this line is used by starport scaffolding # 4 -->
+    </div>
+  </div>
+</template>
+
+<script>
+import * as sp from "@tendermint/vue";
+
+export default {
+  components: { ...sp },
+};
+</script>
