+++ x/issuer/client/rest/query.go
@@ -32,7 +32,7 @@ func getAllIssuersHandlerFn(cliCtx context.CLIContext) http.HandlerFunc {
 		var issuers []types.Issuer
 		for _, kv := range resKVs {
 			issuer := types.Issuer{}
-			cdc.UnmarshalBinaryLengthPrefixed(kv.Value, &issuer)
+			cdc.UnmarshalBinaryBare(kv.Value, &issuer)
 			issuers = append(issuers, issuer)
 
 		}
